Refactoring codegen, tests, cekit

+++ hack/client-gen.sh
@@ -17,6 +17,7 @@ set -o pipefail
 set -o errtrace
 
 location=$(dirname "$0")
+# always consider to upgrade this version when upgrading Operator SDK or k8s/api
 generator_version=v0.21.1
 # in the future this can be an array
 kogito_api_version=v1beta1
@@ -50,10 +51,12 @@ gen_client() {
     --go-header-file=../hack/boilerplate.go.txt \
     --input=/"${kogito_api_version}" \
     --input-base=github.com/kiegroup/kogito-operator/api \
+    --input-dirs=github.com/kiegroup/kogito-operator/api/${kogito_api_version} \
     --clientset-name "versioned" \
     --output-base=. \
     --output-package=github.com/kiegroup/kogito-operator/client/clientset
   error_handler $? "Failed to generate clientset"
+  move_gen_code
 }
 
 gen_listers() {
@@ -69,6 +72,7 @@ gen_informers() {
   go run k8s.io/code-generator/cmd/informer-gen \
     --go-header-file=../hack/boilerplate.go.txt \
     --versioned-clientset-package=github.com/kiegroup/kogito-operator/client/clientset/versioned \
+    --internal-clientset-package=github.com/kiegroup/kogito-operator/client/clientset/versioned \
     --listers-package=github.com/kiegroup/kogito-operator/client/listers \
     --input-dirs=github.com/kiegroup/kogito-operator/api/${kogito_api_version} \
     --output-base=. \
@@ -79,10 +83,14 @@ gen_informers() {
 move_gen_code() {
   mv github.com/kiegroup/kogito-operator/client/* .
   error_handler $? "Failed to move generated code to /client directory"
-  mv clientset/versioned/typed/${kogito_api_version}/_client.go clientset/versioned/typed/${kogito_api_version}/kogito_client.go
-  error_handler $? "Failed to rename _client.go to kogito_client.go"
-  mv clientset/versioned/typed/${kogito_api_version}/fake/fake__client.go clientset/versioned/typed/${kogito_api_version}/fake/fake_kogito_client.go
-  error_handler $? "Failed to rename fake__client.go to fake_kogito_client.go"
+  if [ ! -f "clientset/versioned/typed/${kogito_api_version}/kogito_client.go" ]; then
+    mv clientset/versioned/typed/${kogito_api_version}/_client.go clientset/versioned/typed/${kogito_api_version}/kogito_client.go
+    error_handler $? "Failed to rename _client.go to kogito_client.go"
+  fi
+  if [ ! -f "clientset/versioned/typed/${kogito_api_version}/fake/fake_kogito_client.go" ]; then
+    mv clientset/versioned/typed/${kogito_api_version}/fake/fake__client.go clientset/versioned/typed/${kogito_api_version}/fake/fake_kogito_client.go
+    error_handler $? "Failed to rename fake__client.go to fake_kogito_client.go"
+  fi
 }
 
 set_up
